test(frontend): refine orchestrator operations test helpers
