+++ packages/frontend/src/state/operations/agent-orchestrator/use-agent-orchestrator-operations.test-helpers.ts
@@ -1,0 +1,37 @@
+export { OpencodeSdkAdapter } from "@openducktor/adapters-opencode-sdk";
+export { OPENCODE_RUNTIME_DESCRIPTOR } from "@openducktor/contracts";
+export { toast } from "sonner";
+export { clearAppQueryClient } from "@/lib/query-client";
+export { sessionMessagesToArray } from "@/test-utils/session-message-test-helpers";
+export { host } from "../shared/host";
+export { createSessionMessagesState } from "./support/messages";
+export { createAgentSessionPresenceSnapshotFixture } from "./test-utils";
+export {
+  type OpencodeSdkAdapterPrototype,
+  opencodeSdkAdapterPrototype,
+  type ReadSessionPresenceInput,
+  setupOrchestratorOperationsTestEnvironment,
+} from "./use-agent-orchestrator-operations.test-environment";
+export {
+  BUILD_SELECTION,
+  buildBootstrapFixture,
+  createUnavailableBuildTaskFixture,
+  createWorktreeRuntimeFixture,
+  persistedBuildSessionFixture,
+  persistedSessionFixture,
+  taskFixture,
+  taskFixture2,
+  taskFixture2WithPersistedBuildSession,
+  taskFixtureWithPersistedBuildSession,
+} from "./use-agent-orchestrator-operations.test-fixtures";
+export {
+  createHookHarness,
+  createTestDependencies,
+  type OrchestratorDependencies,
+} from "./use-agent-orchestrator-operations.test-harness";
+export {
+  createDeferred,
+  createPatchScope,
+  runOrchestratorOperationTest,
+  runWithPatchScope,
+} from "./use-agent-orchestrator-operations.test-support";
